extract result extensions
