+++ Homework/TagsCloud.Visualization/WordsParser/WordsParser.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
+using TagsCloud.Visualization.Extensions;
 using TagsCloud.Visualization.Utils;
 using TagsCloud.Visualization.WordsFilter;
